+++ resources/views/default/column/lists.blade.php
@@ -1,8 +1,26 @@
 <div {!! $attributes !!}>
   @if ($visibled)
-    @foreach ($values as $value)
-      <span class="badge table-badge" v-pre>{!! $value !!}</span>
-    @endforeach
+    @if ($maxLists && count($values) > $maxLists)
+      @php
+        $count = 0;
+        $more = count($values) - $maxLists;
+      @endphp
+
+      @foreach ($values as $value)
+        @php
+          if(++$count > $maxLists) break;
+        @endphp
+        <span class="badge table-badge" v-pre>{!! $value !!}</span>
+      @endforeach
+        <span class="badge text-muted">{{ trans('sleeping_owl::lang.select.more', ['count' => $more]) }}</span>
+
+    @else
+      @foreach ($values as $value)
+        <span class="badge table-badge" v-pre>{!! $value !!}</span>
+      @endforeach
+    @endif
+
+
     {!! $append !!}
 
     @if($small)